content_analyzer/requirements.txt
# Core dependencies - minimal set
Flask==2.3.2
pdfplumber==0.10.2
python-docx==1.1.2
Pillow==12.0.0
pytesseract==0.3.10
reportlab==4.0.4

# Web server for production
gunicorn==20.1.0

# Google Cloud dependencies for Vertex AI
google-auth==2.22.0
google-api-python-client==2.95.0
google-cloud-aiplatform==1.26.1
google-generativeai
python-dotenv